requirements.txt
# Core ETL & Web Scraping
pandas~=2.2.0
requests~=2.32.0
beautifulsoup4~=4.13.0

# Database (PostgreSQL)
sqlalchemy~=2.0.0
psycopg2-binary~=2.9.0

# Google Sheets API
google-auth~=2.39.0
google-api-python-client~=2.168.0
gspread~=5.12.0 
googleapis-common-protos~=1.60.0
proto-plus~=1.23.0

# Testing
pytest~=8.3.0
pytest-cov~=5.0.0
pytest-mock~=3.14.0

# Optional: Untuk .env file
python-dotenv~=1.1.0